move pool_stats_collector
