+++ ydb/library/pool_stats_collector/pool_stats_collector.h
@@ -1,5 +1,6 @@
 #pragma once
-#include "defs.h"
+
+#include <ydb/core/base/defs.h>
 #include <ydb/library/actors/core/actorsystem.h>
 #include <library/cpp/monlib/dynamic_counters/counters.h>
